OneDrive - UA Little Rock/Final Grad Project/AdFilteringEngine_FinalDemo_defense/requirements.txt
PyQt5
PyQtWebEngine
pandas
tensorflow
keras.preprocessing
